refactor: 重構 RecipeDraftVideo 元件，簡化結構並將商業邏輯移至工具函數和自定義 hook，提升可讀性與維護性。
