+++ src/components/pages/RecipeDraftVideo/BreadcrumbNavigation.tsx
@@ -1,0 +1,39 @@
+import {
+  Breadcrumb,
+  BreadcrumbList,
+  BreadcrumbItem,
+  BreadcrumbLink,
+  BreadcrumbPage,
+  BreadcrumbSeparator,
+} from '@/components/ui/breadcrumb';
+
+type BreadcrumbNavigationProps = {
+  userDisplayId: string | null;
+};
+
+/**
+ * 麵包屑導航元件
+ */
+export const BreadcrumbNavigation: React.FC<BreadcrumbNavigationProps> = ({
+  userDisplayId,
+}) => (
+  <div className="p-4">
+    <Breadcrumb>
+      <BreadcrumbList>
+        <BreadcrumbItem>
+          <BreadcrumbLink href="/">首頁</BreadcrumbLink>
+        </BreadcrumbItem>
+        <BreadcrumbSeparator />
+        <BreadcrumbItem>
+          <BreadcrumbLink href={userDisplayId ? `/user/${userDisplayId}` : '/'}>
+            會員中心
+          </BreadcrumbLink>
+        </BreadcrumbItem>
+        <BreadcrumbSeparator />
+        <BreadcrumbItem>
+          <BreadcrumbPage>影片步驟確認</BreadcrumbPage>
+        </BreadcrumbItem>
+      </BreadcrumbList>
+    </Breadcrumb>
+  </div>
+);
